refactor(domain): group vertical slice types in subfolder
