+++ src/core/domain/vertical-slice/VerticalSliceProjection.ts
