+++ tests/core/usecases/ExportJSON/ExportJSONQueryHandler.test.ts
@@ -5,8 +5,8 @@ import { CommandNode } from '../../../../src/core/domain/CommandNode';
 import { ReadModelNode } from '../../../../src/core/domain/ReadModelNode';
 import { PolicyNode } from '../../../../src/core/domain/PolicyNode';
 import { UIScreenNode } from '../../../../src/core/domain/UIScreenNode';
-import { VerticalSliceCollection } from '../../../../src/core/domain/VerticalSliceCollection';
-import { VerticalSlice } from '../../../../src/core/domain/VerticalSlice';
+import { VerticalSliceCollection } from '../../../../src/core/domain/vertical-slice/VerticalSliceCollection';
+import { VerticalSlice } from '../../../../src/core/domain/vertical-slice/VerticalSlice';
 import { Scenario } from '../../../../src/core/domain/Scenario';
 import { type NodeProperties } from '../../../../src/core/domain/NodeProperties';
 import { ExportJSONQuery } from '../../../../src/core/usecases/queries/ExportJSON/ExportJSONQuery';
