+++ tests/core/usecases/ExportMarkdown/ExportMarkdownQueryHandler.test.ts
@@ -5,7 +5,7 @@ import { CommandNode } from '../../../../src/core/domain/CommandNode';
 import { ReadModelNode } from '../../../../src/core/domain/ReadModelNode';
 import { PolicyNode } from '../../../../src/core/domain/PolicyNode';
 import { UIScreenNode } from '../../../../src/core/domain/UIScreenNode';
-import { VerticalSliceCollection } from '../../../../src/core/domain/VerticalSliceCollection';
+import { VerticalSliceCollection } from '../../../../src/core/domain/vertical-slice/VerticalSliceCollection';
 import { ExportMarkdownQuery } from '../../../../src/core/usecases/queries/ExportMarkdown/ExportMarkdownQuery';
 import { ExportMarkdownQueryHandler } from '../../../../src/core/usecases/queries/ExportMarkdown/ExportMarkdownQueryHandler';
