+++ tests/core/usecases/RemoveScenarioFromSlice/RemoveScenarioFromSliceCommandHandler.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
-import { VerticalSliceCollection } from '../../../../src/core/domain/VerticalSliceCollection';
-import { VerticalSlice } from '../../../../src/core/domain/VerticalSlice';
+import { VerticalSliceCollection } from '../../../../src/core/domain/vertical-slice/VerticalSliceCollection';
+import { VerticalSlice } from '../../../../src/core/domain/vertical-slice/VerticalSlice';
 import { Scenario } from '../../../../src/core/domain/Scenario';
 import { RemoveScenarioFromSliceCommand } from '../../../../src/core/usecases/commands/RemoveScenarioFromSlice/RemoveScenarioFromSliceCommand';
 import { RemoveScenarioFromSliceCommandHandler } from '../../../../src/core/usecases/commands/RemoveScenarioFromSlice/RemoveScenarioFromSliceCommandHandler';
